Removed all code that dealt with non-flattened IR
Just realized most of the cases I was trying to account for should never
come up once the IR is flattened and instrumentation begins. I've
removed that code, and kept a backup of it in hg_main_backup.c so it can
be referred to for writing future similar cases.