Added support for the rest of the cases
Well, at least all the parts of the switch statement are filled in. I
still need to do something to instrument operations and actually create
shadow values, but all the moving them around seems to be done. I also
in this commit removed the shadow value copying functions that only
apply to IR that hasn't been flattened.