Added another layer of indirection to shadow vals
One of the main problems we ran into with FpDebug was the fact that,
since it only had a single layer of indirection for shadow values, you
could only store one shadow value at each logical value
location (memory, temps, guest registers). Since the locations could be
up to 128 bits wide in VEX IR, it's actually possible to store multiple
shadow values in a single location in the IR, or just move a 64-bit
value into the low order bits of a 128-bit location, and then out
again. While it doesn't really make sense to do these things, somewhere
along the pipeline from c code, compiled to x86, and then decompiled to
VEX IR, these things end up cropping up, and so not being able to handle
them means you'll "lose" shadow values that go through these weird
partial locations. Instead of making this mistake, we're adding another
layer of indirection, called shadow locations. A shadow location can be
stored in memory, in a guest register, or in a temporary, just like the
old shadow values. It will only be created when it has an actual shadow
value to be stored there, as in we think that what is stored there is a
float. Shadow locations will be able to each store an arbitrary but
fixed number of shadow values.