Support a weird read
Support the case where someone tries to read a 128 bit value from a
location where they'd previously placed a 64 bit float. Also tries to
pull in the next 64 bits of memory as the second half of the 128-bit, if
it's a value we have shadowed. Otherwise, just leave the shadow blank.