Made const arguments uniformly moved to temps
We did this in one place before, now we do it everywhere, and it's
abstracted into a function. This is technically uneccessary, since the
constants will not already have a shadow value, so they'll just get
picked up from a load into memory, but we might want it later and it
makes things cleaner.