Moved responsibility for initing ASTs to later
Specifically, leaf AST's were being initialized on the creation of every
shadow location, and then immediately overwritten with branch AST's when
they were used for the result of an operation. Now, we don't initialize
them at all on creation, and rely on the sites of creation to know what
they'll be used for, and initialize them properly.