Generate C macros for libm funcs via python
In the last commit I abstracted all the code that cared about which libm
functions we supported into a single file of macros, and now I'm
generating that macro file using a python script.

This means that you can add a single line to the python script (look
near the bottom for examples), to add a new supported op to herbgrind,
and everything should work just fine. The errors if you mess it up,
unfortunately, might be inscrutiable. Oh well.

So, we now git ignore the generated header, since it's not really
source-level anymore. I added the python script to the repo and makefile
list of headers, and in the makefile we run it to generate the header
every time we compile.

This should make it pretty much trivial to add all the libm functions we want.