Added a benchmark which exposes the bug
var-test.c doesn't expose the bug because it first creates the "shorter"
ast, then generalizes it with the longer one. This one does it the other
way around, and triggers the assertion which reveals the broken
assumption. I'm keeping the old one around because it's good to test it
both ways, and naming this one var-test-2