Fixed band-aid?
I did a little more investigation while a gromacs run was going, and it
looks like the only time we had a problem (at least in triangle), what
was happening was we had an operation which took a multi-channel float
register and did something with the low order float, and valgrind seemed
to be getting a bit confused. For instance, an operation like Mul64F0x2,
which takes two arguments of two 64-bit floats each, and multiplies the
low order floats together to get the low order result, and copies the
high-order result from the first argument. Sometimes, valgrind would
type a temporary with only a single 64-bit float, move a 64-bit float
into it, and then change the type to a dual channel 64-bit float
register before doing the operation. Which doesn't matter too much,
since the operation is only on the low order thing which existed
anyway. So, I just expand the shadow location in that case, and it all
seems to go through.