Fixed simulating normal precision
Our previous simulation of normal precision, both for local error and
for mathwrap stuff, was actually too accurate. For some reason I thought
that the mpfr precision you set was the size of the whole float you want
to simulate. But actually, you're setting the size of the mantissa, and
the exponent is fixed at 11 bits. So I was giving "normal" precision
calculations 64 bits of mantissa, and 75 bits total. I fixed this, now
we simulate 64-bit stuff correctly. This should make local error more
accurate, as well as preventing mathwrapping from changing the client
semantics (oops!).