Add basic symbolic ops infrastructure
This includes the code for symbolic ops, with two noticable omissions:
the generalize an existing expr case is omitted, as well as the
getEquivalences code. This means we can perfectly create and print a
symbolic op that is a direct translation from a single concrete
expression, but we will fail an assert when multiple concrete
expressions are generalized together. The net result is that this
behaves exactly like the previous commit, except it will assert fail
more often, and has a lot more support code.