Restructuring variable mapping and structure generalization
The new idea is that there are inherent inefficiencies in both. The old
structure generalization is checking for divergences at every node, but
we can actually use the superblock boundries to figure out the possible
place where the tree could diverge, called "graft-points", and only
check those. For variable mappings, we were computing the entire
matching set on the new tree and attempting to merge it with the entire
matching set on the old one, when most nodes in practice will have no
equivalence information. So now we will only check the ones that do.