Rename isFloatType to isFloatIRType, and add new isFloatType
The function that used to be called "isFloatType" took in IRTypes, and
determined whether they could hold a float. This is now renamed to
isFloatIRType to make more clear what kind of data it operates on. The
new function called isFloatType takes in ValueTypes, and determines
whether they are consistent with a float value. The type system should
catch any accidental uses of the new method.