Initialize shadows of partially initialized temps when needed
With the new block based storage systems, you can create a 2-block
temp that's initialized, then store it somewhere, pull 4 blocks from 2
blocks before it to make a temp, and end up with a temp with it's
first two values uninitialized. When you try and operate on this the
system would get confused that the temp is non null, but the shadow value it
wants isn't. This commit adds some code to create the shadow value
from the client value in this case.